chore : renama package from autoscheduling -> scheduling as per suggestion
