+++ pkg/manager/scheduling/baremetal.go
@@ -1,4 +1,4 @@
-package autoscheduling
+package scheduling
 
 import (
 	"fmt"
